Refactor troubleshooting challenge description
This started out as a patch to add Kubernetes to the list of challenges,
but that required refactoring our wording around "we don't ask
application questions".  While refactoring that wording, I noticed some
"zoom" capitalization issues, and the next thing you know, I've
rephrased a large part of the challenge.

This also adds a note that it is ok to install troubleshooting packages,
as strong candidates commonly do this.