frontend: Replace mention of Airbnb coding style
Grepping through the project and git history, it doesn't look like we ever
used eslint-config-airbnb. Having spent a month in the project, it seems
that our linter rules are not as strict as Airbnb's JS style guide, so let's
replace mentions of Airbnb coding style with stuff that we actually use.